toolkit/components/viewsource/test/test_428653.xul

<?xml-stylesheet href="chrome://global/skin/" type="text/css"?> 
<?xml-stylesheet
  href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="View Source Test (bug 428653)"
        xmlns:html="http://www.w3.org/1999/xhtml"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
   src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
   
  <html:body/>

  <browser id="content" type="content-primary" name="content" src="about:blank" flex="1"
           disablehistory="true" context="viewSourceContextMenu"/>

  <script type="application/javascript">
  <![CDATA[
  /*
  Test that we can't call the content browser's document.open() over Xrays.
  See the security checks in nsHTMLDocument::Open, which make sure that the
  entry global's principal matches that of the document.
  */
  SimpleTest.waitForExplicitFinish();
  
  addLoadEvent(function testDocumentOpen() {
    var browser = document.getElementById("content");
    ok(browser, "got browser");
    var doc = browser.contentDocument;
    ok(doc, "got content document");

    var opened = false;
    try {
      doc.open("text/html", "replace");
      opened = true;
    } catch (e) {
      is(e.name, "SecurityError", "Unxpected exception")
    }
    is(opened, false, "Shouldn't have opened document");

    doc.wrappedJSObject.open("text/html", "replace");
    ok(true, "Should be able to open document via Xray Waiver");

    SimpleTest.finish();
  });
  ]]>
  </script>
</window>
